+++ demisto_sdk/commands/common/tests/update_id_set_test.py
@@ -12,8 +12,9 @@ from demisto_sdk.commands.common.update_id_set import (
     find_duplicates, get_classifier_data, get_fields_by_script_argument,
     get_incident_field_data, get_incident_fields_by_playbook_input,
     get_incident_type_data, get_indicator_type_data, get_integration_data,
-    get_layout_data, get_mapper_data, get_playbook_data, get_script_data,
-    get_values_for_keys_recursively, has_duplicate, re_create_id_set)
+    get_layout_data, get_layoutscontainer_data, get_mapper_data,
+    get_playbook_data, get_script_data, get_values_for_keys_recursively,
+    has_duplicate, re_create_id_set)
 from demisto_sdk.commands.create_id_set.create_id_set import IDSetCreator
 from TestSuite.utils import IsEqualFunctions
 
@@ -377,6 +378,52 @@ class TestPlaybooks:
         assert 'incident_fields' not in result.keys()
         assert 'indicator_fields' not in result.keys()
 
+    @staticmethod
+    def test_get_playbook_data_bad_graph():
+        """
+        Given
+            - A playbook file called playbook-invalid-bad-graph.yml:
+                - task 1 point to non-existing task
+                - task 2 is not connected
+
+        When
+            - parsing playbook files
+
+        Then
+            - parsing flow graph from file successfully (only tasks 0 and 1 will be in the graph)
+        """
+        test_file_path = os.path.join(TESTS_DIR, 'test_files', 'playbook-invalid-bad-graph.yml')
+        result = get_playbook_data(test_file_path)
+        playbook_data = result.get('InvalidPlaybook-BadGraph', {})
+        assert playbook_data.get('name') == 'InvalidPlaybook-BadGraph'
+        assert playbook_data.get('command_to_integration', {}).get('ip') == ''
+        assert playbook_data.get('command_to_integration', {}).get('domain') == ''
+        assert 'domain' in playbook_data.get('skippable_tasks', [])
+        assert 'ip' not in playbook_data.get('skippable_tasks', [])
+
+    @staticmethod
+    def test_get_playbook_data_bad_graph_2():
+        """
+        Given
+            - A playbook file called playbook-invalid-bad-graph_2.yml:
+                - starttaskid=5 but task 5 does not exist
+
+        When
+            - parsing playbook files
+
+        Then
+            - parsing flow graph from file successfully (no actual tasks will be in the graph)
+        """
+        test_file_path = os.path.join(TESTS_DIR, 'test_files', 'playbook-invalid-bad-graph_2.yml')
+        result = get_playbook_data(test_file_path)
+        playbook_data = result.get('InvalidPlaybook-BadGraph', {})
+        assert playbook_data.get('name') == 'InvalidPlaybook-BadGraph'
+        assert playbook_data.get('command_to_integration', {}).get('ip') == ''
+        assert playbook_data.get('command_to_integration', {}).get('domain') == ''
+        # domain task is marked as skippable so it will be included regardless to the graph.
+        assert 'domain' in playbook_data.get('skippable_tasks', [])
+        assert len(playbook_data.get('skippable_tasks', [])) == 1
+
 
 class TestLayouts:
     @staticmethod
@@ -426,6 +473,30 @@ class TestLayouts:
         assert 'typeID' in result.keys()
         assert 'incident_and_indicator_types' in result.keys()
         assert 'incident_and_indicator_fields' not in result.keys()
+
+    @staticmethod
+    def test_get_layoutscontainer_data():
+        """
+        Given
+            - A layoutscontainer file called layoutscontainer-to-test.json
+
+        When
+            - parsing layoutscontainer files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/layoutscontainer-to-test.json'
+        result = get_layoutscontainer_data(test_dir)
+        result = result.get('layouts_container_test')
+        assert 'detailsV2' in result.keys()
+        assert 'name' in result.keys()
+        assert 'group' in result.keys()
+        assert 'fromversion' in result.keys()
+        assert 'toversion' in result.keys()
+        assert 'file_path' in result.keys()
+        assert 'incident_and_indicator_types' in result.keys()
+        assert 'incident_and_indicator_fields' in result.keys()
 
 
 class TestIncidentFields:
@@ -789,7 +860,7 @@ class TestGenericFunctions:
             -  Finding all dependent incident fields in the input
         """
 
-        result = get_incident_fields_by_playbook_input(input=playbook_input.get('value'))
+        result = get_incident_fields_by_playbook_input(playbook_input.get('value'))
         if are_there_incident_fields:
             assert "field_name" in result
         else: