+++ demisto_sdk/commands/validate/file_validator.py
@@ -27,13 +27,14 @@ from demisto_sdk.commands.common.constants import (
     IMAGE_REGEX, INTEGRATION_REGXES, JSON_ALL_CLASSIFIER_REGEXES,
     JSON_ALL_CLASSIFIER_REGEXES_5_9_9, JSON_ALL_DASHBOARDS_REGEXES,
     JSON_ALL_INCIDENT_TYPES_REGEXES, JSON_ALL_INDICATOR_TYPES_REGEXES,
-    JSON_ALL_LAYOUT_REGEXES, JSON_ALL_MAPPER_REGEXES,
-    JSON_INDICATOR_AND_INCIDENT_FIELDS, KNOWN_FILE_STATUSES,
-    OLD_YML_FORMAT_FILE, PACKAGE_SCRIPTS_REGEXES, PACKS_DIR,
-    PACKS_INTEGRATION_NON_SPLIT_YML_REGEX, PACKS_PACK_IGNORE_FILE_NAME,
-    PACKS_RELEASE_NOTES_REGEX, PACKS_SCRIPT_NON_SPLIT_YML_REGEX,
-    PLAYBOOK_REGEX, PLAYBOOKS_REGEXES_LIST, SCHEMA_REGEX, TEST_PLAYBOOK_REGEX,
-    YML_ALL_SCRIPTS_REGEXES, YML_INTEGRATION_REGEXES)
+    JSON_ALL_LAYOUT_REGEXES, JSON_ALL_LAYOUTS_CONTAINER_REGEXES,
+    JSON_ALL_MAPPER_REGEXES, JSON_INDICATOR_AND_INCIDENT_FIELDS,
+    KNOWN_FILE_STATUSES, OLD_YML_FORMAT_FILE, PACKAGE_SCRIPTS_REGEXES,
+    PACKS_DIR, PACKS_INTEGRATION_NON_SPLIT_YML_REGEX,
+    PACKS_PACK_IGNORE_FILE_NAME, PACKS_RELEASE_NOTES_REGEX,
+    PACKS_SCRIPT_NON_SPLIT_YML_REGEX, PLAYBOOK_REGEX, PLAYBOOKS_REGEXES_LIST,
+    SCHEMA_REGEX, TEST_PLAYBOOK_REGEX, YML_ALL_SCRIPTS_REGEXES,
+    YML_INTEGRATION_REGEXES)
 from demisto_sdk.commands.common.errors import (ALLOWED_IGNORE_ERRORS,
                                                 ERROR_CODE,
                                                 FOUND_FILES_AND_ERRORS,
@@ -56,7 +57,8 @@ from demisto_sdk.commands.common.hook_validations.incident_type import \
     IncidentTypeValidator
 from demisto_sdk.commands.common.hook_validations.integration import \
     IntegrationValidator
-from demisto_sdk.commands.common.hook_validations.layout import LayoutValidator
+from demisto_sdk.commands.common.hook_validations.layout import (
+    LayoutsContainerValidator, LayoutValidator)
 from demisto_sdk.commands.common.hook_validations.mapper import MapperValidator
 from demisto_sdk.commands.common.hook_validations.pack_unique_files import \
     PackUniqueFilesValidator
@@ -478,6 +480,12 @@ class FilesValidator:
                 reputation_validator = ReputationValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                            print_as_warnings=self.print_ignored_errors)
                 if not reputation_validator.is_valid_file(validate_rn=True):
+                    self._is_valid = False
+
+            elif checked_type(file_path, JSON_ALL_LAYOUTS_CONTAINER_REGEXES):
+                layout_validator = LayoutsContainerValidator(structure_validator, ignored_errors=ignored_errors_list,
+                                                             print_as_warnings=self.print_ignored_errors)
+                if not layout_validator.is_valid_layout(validate_rn=True):
                     self._is_valid = False
 
             elif checked_type(file_path, JSON_ALL_LAYOUT_REGEXES):