+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Fixed an issue where release notes were not checked correctly in **validate** command.
 * Fixed an issue in the **create-id-set** command where optional playbook tasks were not taken into consideration.
 * Added a prompt to the `demisto-sdk update-release-notes` command to prompt users to commit changes before running the release notes command.
+* Added support to `layoutscontainer` in **validate** command.
 
 #### 1.1.5
 * Fixed an issue in **find-dependencies** command.